requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
xgboost==2.0.3
python-telegram-bot==20.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
openai==2.15.0
google-generativeai>=0.8.0
groq>=0.9.0
beautifulsoup4==4.12.3
requests==2.31.0
